Expose static factory methods for FontSpec
